feat(apimock): automatic fallbacks
ApiMock can now activate automatically on request errors. So it’ll try
to access the API and if it fails then it will try to do a rerouted
mock request instead. The flag is `auto` and can be set globally in
browser url `?apiMock=auto` or locally on $http request (`$http.get({
url: ‘…’, apiMock: ‘auto’ })`).