refactor(apiMock): major rewrite
Rewrote most of the code for a cleaner interface that supports more
potential commands. Previous version only worked will with auto on/off.

BREAKING CHANGE: Interface for apiMock service has changed.

You shouldn’t be using apiMock object anymore. There’s currently not a
good way to detect if mocking is enabled as that can change on
individual http-requests anyway.

We’re considering printing to $log to notify when mocking is occuring.
Suggestions are welcome.