fix(travis): don't run SauceLabs tests on PR's
The fix from #32 made it so that the build always passes, which wasn’t
the idea. This was happened on #33 which triggered two Travis builds:
one as PR which didn’t run SL so it passed, and another as a regular
Push that had failing SL tests (IE8 and trailing commas ftw) but still
passed.

So the conditional is greatly simplified. If it’s a PR just pass it,
otherwise run the SL tests. Don’t know why the Travis docs had such a
backwards workaround sample.

Fixes #31