chore(grunt): commit nuget file
The “npm-publish” task fails because the workspace isn’t clean. The
nuget task creates a new .nupkg file which isn’t tracked in git. Adding
a “gitadd” task to include it so the publish task can complete
successfully.

I haven’t put the .nupkg filepath in a variable because I’m not sure if
the “yeoman.version” variable will update and propagate properly.

Fixes #54